+++ CHANGES.md
@@ -65,6 +65,7 @@ To be released.
      -  Added `LookupWebFingerOptions.tracerProvider` option.
      -  Added `LookupObjectOptions.tracerProvider` option.
      -  Added `GetActorHandleOptions.tracerProvider` option.
+     -  Added `VerifyRequestOptions.tracerProvider` option.
 
  -  Added `@fedify/fedify/x/sveltekit` module for integrating with [SvelteKit]
     hook.  [[#171], [#183] by Jiyu Park]